% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{getAllTimeRecordVsFranchise}
\alias{getAllTimeRecordVsFranchise}
\title{Get Franchise Head to Head record}
\usage{
getAllTimeRecordVsFranchise(franchiseID = NULL)
}
\arguments{
\item{franchiseID}{Optional franchise ID to filter results.}
}
\value{
Head to head records for every franchise.
}
\description{
Get Franchise Head to Head record
}
