% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhlRapi.R
\name{getAPI}
\alias{getAPI}
\title{Base getAPI call}
\usage{
getAPI(query, modifiers = NULL)
}
\arguments{
\item{query}{base query}

\item{modifiers}{modifiers}
}
\value{
response parsed by jsonlite
}
\description{
Base API calling code for internal use. Use specific endpoints to build queries in recognized fasion.
}
\keyword{internal}
