% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{getAttendance}
\alias{getAttendance}
\title{Get Attendance Records}
\usage{
getAttendance(season = NULL)
}
\arguments{
\item{season}{Optional, to pull attendance from given season. Format: 20172018}
}
\value{
season attendance records from the API
}
\description{
Get Attendance Records
}
\examples{
getAttendance()
getAttendance(20152016)
}
