% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conferences.R
\name{getConference}
\alias{getConference}
\title{Get Conference}
\usage{
getConference(conference = NULL)
}
\arguments{
\item{conference}{Optional, get specific conference information. Eastern = 6, Western = 5, World cup of Hockey = 7}
}
\value{
The API output of conferences
}
\description{
Get the NHL conferences, including historical conferences. Or, get one single conference
}
\examples{
#See information on the NHL conferences:
conferences <- getConference()
}
