% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhlRapi.R
\name{getCopyright}
\alias{getCopyright}
\title{Get Copyright Statement}
\usage{
getCopyright()
}
\value{
The copyright statement from the API calls
}
\description{
return the copyright statement from the API calls
}
