% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions.R
\name{getDivision}
\alias{getDivision}
\title{Get Division}
\usage{
getDivision(division = NULL)
}
\arguments{
\item{division}{Optional, get specific division information. Pacific = 15, Central = 16, Atlantic = 17, Metropolitan = 18.}
}
\value{
The API output of divisions
}
\description{
Get the NHL divisions, including historical divisions. Or, get one single division
}
\examples{
#See information on NHL Divisions:
divisions<-getDivision()
}
