% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drafts.R
\name{getDraft}
\alias{getDraft}
\title{Get Draft}
\usage{
getDraft(year = NULL)
}
\arguments{
\item{year}{Optional, get specific draft year information. If not set, returns current year's draft. First draft was 1963, so any year prior will result in error}
}
\value{
The API output of draft
}
\description{
Get round by round data for draft year(s).
}
\examples{
#Get draft data for 2002 entry draft
getDraft(2002)
}
