% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drafts.R
\name{getDraftRecords}
\alias{getDraftRecords}
\title{getDraftRecords}
\usage{
getDraftRecords(firstName = NULL, lastName = NULL, draftYear = NULL,
  overallPickNumber = NULL, weight = NULL, height = NULL,
  position = NULL, roundNumber = NULL, draftTeamID = NULL,
  amateurLeague = NULL, country = NULL)
}
\arguments{
\item{firstName}{Optional, drafted player first name}

\item{lastName}{Optional, drafted player last name}

\item{draftYear}{Optional, drafted player drafted year}

\item{overallPickNumber}{Optional, drafted player overall pick number}

\item{weight}{Optional, drafted player weight}

\item{height}{Optional, drafted player height}

\item{position}{Optional, drafted player position (LW, C, RW, D, G, F)}

\item{roundNumber}{Optional, round in which player was drafterd}

\item{draftTeamID}{Optional, drafting TEAM ID}

\item{amateurLeague}{Optional, drafted player previous amateur league. Many NA values in API database}

\item{country}{Optional, drafted player country}
}
\value{
API results from the drafts endpoint, filtered using any parameters
}
\description{
Pull up information on specific draft picks. Many filters to chose from.
}
\examples{
#See all G drafted in 2006:
getDraftRecords(position="G", draftYear = 2006)
}
