% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{getFranchiseDetail}
\alias{getFranchiseDetail}
\title{Get Franchise Detailed Information}
\usage{
getFranchiseDetail(teamID = NULL)
}
\arguments{
\item{teamID}{Optional team ID to filter results.}
}
\value{
detailed information for evey franchise, including captains, coaches, GMs, retired numbers, etc.
}
\description{
Get Franchise Detailed Information
}
