% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{getFranchiseGoalieRecords}
\alias{getFranchiseGoalieRecords}
\title{Get Franchise Goalie Record Stats}
\usage{
getFranchiseGoalieRecords(franchiseID = NULL)
}
\arguments{
\item{franchiseID}{Optional franchise ID to filter results.}
}
\value{
Season record stats for every franchise.
}
\description{
Get Franchise Goalie Record Stats
}
