% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{getFranchiseRecords}
\alias{getFranchiseRecords}
\title{Get Franchise Record Stats}
\usage{
getFranchiseRecords(franchiseID = NULL, franchiseName = NULL)
}
\arguments{
\item{franchiseID}{Optional franchise ID to filter results.}

\item{franchiseName}{Optional franchise name to filter results.}
}
\value{
Season record stats for every (filtered) franchise.
}
\description{
Retrieves franchise records (best/worst seasons, most goals in a game, etc).
}
\examples{
#See the records for franchise 23:
records<-getFranchiseRecords(franchiseID = 23)

#Or, knowing that franchise 23 is New Jersey Devils:
records<-getFranchiseRecords(franchiseName = "New Jersey Devils")
}
