% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{getFranchiseTeamTotal}
\alias{getFranchiseTeamTotal}
\title{Get Franchise Team Total Stats}
\usage{
getFranchiseTeamTotal(franchiseID = NULL)
}
\arguments{
\item{franchiseID}{Optional franchise ID to filter results.}
}
\value{
total stats for every franchise.
}
\description{
Get Franchise Team Total Stats
}
