% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{getGameLinescore}
\alias{getGameLinescore}
\title{Get Game Linescore}
\usage{
getGameLinescore(game)
}
\arguments{
\item{game}{Required, get specific game boxscore.}
}
\value{
The linescore output of the API
}
\description{
Get a NHL game linescore. Includes goals, shots on goal, powerplay and goalie pulled status, number of skaters and shootout information if applicable
}
