% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestones.R
\name{getMilestone}
\alias{getMilestone}
\title{Get Milestones Records}
\usage{
getMilestone(milestone)
}
\arguments{
\item{milestone}{Required, one of:
\itemize{
  \item \code{1000-point-career}
  \item \code{500-goal-career}
  \item \code{100-point-season}
  \item \code{50-goal-season}
  \item \code{5-goal-game}
}}
}
\value{
milestones from the API of a requested type
}
\description{
The NHL tracks a few milestones of public interest. This allows calling up the milestones achieved by plyaers. See the milestone parameter for milestones tracked.
}
\examples{
#see all 5 goal games:
five_goal_games <- getMilestone('5-goal-game')
}
