% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{getPlayerRecord}
\alias{getPlayerRecord}
\title{Get Player Information}
\usage{
getPlayerRecord(firstName = NULL, lastName = NULL, middleName = NULL,
  position = NULL, playerID = NULL, nationality = NULL)
}
\arguments{
\item{firstName}{Optional, filter by first name}

\item{lastName}{Optional, filter by last name}

\item{middleName}{Optional, filter by middle name}

\item{position}{Optional, filter by position}

\item{playerID}{Optional, numeric player id}

\item{nationality}{Optional, filter by nationality}
}
\value{
Player info from API, filtered with selected filters
}
\description{
Returns lots of metadata on players. Search by ID, first, middle & last name, position, country,
}
\examples{
#Pull up Dave Andreychuk's information
getPlayerRecord(firstName = 'Dave', lastName = 'Andreychuk')
}
