% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{getPlayers}
\alias{getPlayers}
\title{Get Players}
\usage{
getPlayers(player = NULL)
}
\arguments{
\item{player}{Mandatory, get a specific player's information.}
}
\value{
The API output of player
}
\description{
Get NHL players basic info & stats, including historical players.
}
\examples{
#See info on Sidney Crosby (ID = 8471675)
crosby <- getPlayers(8471675)
}
