% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{getPlayersByTeam}
\alias{getPlayersByTeam}
\title{Get Players per team}
\usage{
getPlayersByTeam(team)
}
\arguments{
\item{team}{Team id}
}
\value{
the players (from records api) on the team
}
\description{
Get Players per team
}
