% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{getPlayerStats}
\alias{getPlayerStats}
\title{Get Player's Stats}
\usage{
getPlayerStats(player, stat = NULL, season = NULL)
}
\arguments{
\item{player}{The player ID to get stats}

\item{stat}{The speific stats to get. Many options exist:
- statsSingleSeason - Also provide a season
- homeAndAway - Split by home and away games
- winLoss - Split stats by win/loss/tie outcome
- byMonth -
- byDayOfWeek -
- vsDivision - split stats by division of opponent
- vsConference - split stats by conference of opponent
- vsTeam - split stats by opponent team
- gameLog - provides game by game log of stats
- regularSeasonStatRankings - provides ranking of each stat in season
- goalsByGameSituation
- onPaceRegularSeason - only works for current season, shows pace projected
@param season a season, if required or desired, for querying stats}
}
\value{
The API output of player
}
\description{
get specific stats for a provided player
}
