% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{getPlayerStatTypes}
\alias{getPlayerStatTypes}
\title{Get Player Stat Types}
\usage{
getPlayerStatTypes()
}
\value{
a list of player stat types to call with \code{\link{getPlayerStats}()}
}
\description{
Only certain stat types are accepted for players. This returns the full valid list.
}
\examples{
#See the possible stat types:
getPlayerStatTypes()
}
