% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playoffs.R
\name{getPlayoffDetails}
\alias{getPlayoffDetails}
\title{Get Playoff Details}
\usage{
getPlayoffDetails(season = NULL, series = NULL)
}
\arguments{
\item{season}{optional, the season of playoff info you wish to get}

\item{series}{optional, either series number (1-4) or name ('Conference Quarterfinals', 'Conference Semifinals', 'Conference Finals', 'Stanley Cup Final').}
}
\value{
a list of all playoffs result matching provided season and series
}
\description{
Get Playoff Details
}
\examples{
#Get all playoff series in 20172018
playoffs_1718<-getPlayoffDetails(season = 20172018)

#Get all Stanley Cup Finals series:
playoffs_finals<-getPlayoffDetails(series = 'Stanley Cup Final')

#alternativly, with a series code:
playoffs_finals<-getPlayoffDetails(series = 4)
}
