% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{getPlayoffRecordVsFranchise}
\alias{getPlayoffRecordVsFranchise}
\title{Get Franchise Head to Head playoff record}
\usage{
getPlayoffRecordVsFranchise(franchiseID = NULL, franchiseName = NULL,
  opponentName = NULL, opponentID = NULL)
}
\arguments{
\item{franchiseID}{Optional franchise ID to filter results.}

\item{franchiseName}{Optional franchise name to filter results.}

\item{opponentName}{Optional opponent name to filter results.}

\item{opponentID}{Optional opponent franchise ID to filter results.}
}
\value{
Head to head records for every franchise' playoff games.
}
\description{
Get Franchise Head to Head playoff record
}
