% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drafts.R
\name{getProspect}
\alias{getProspect}
\title{Get Prospect}
\usage{
getProspect(prospect = NULL)
}
\arguments{
\item{prospect}{Optional, get specific prospect information.}
}
\value{
The API output of prospects
}
\description{
Get the NHL draft prospects for the current year, or get one single prospect
}
