% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{getRecordsList}
\alias{getRecordsList}
\title{Get Records List}
\usage{
getRecordsList(recordKey = NULL)
}
\arguments{
\item{recordKey}{optional, filter result by record key}
}
\value{
a list of all records maintained by the NHL. The return is the API URL to call for the current record result.
}
\description{
The NHL maintains a sizeable list of records. Mosts, longests, fewests, youngests, etc.
}
\examples{
records <- getRecordsList(recordKey = "most-shots-one-game")
}
