% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{getSchedule}
\alias{getSchedule}
\title{Get Schedule}
\usage{
getSchedule(team = NULL, startDate = NULL, endDate = NULL,
  scheduleType = NULL)
}
\arguments{
\item{team}{Specific team's schedule to retrieve. optional, numeric}

\item{startDate}{A single date or a date range may be entered, optional, must be coercable with as.Date()}

\item{endDate}{If a date range is requested, the end date. For a single date, this may be left out.}

\item{scheduleType}{additional schedule information may be requested:
\describe{
  \item{\code{broadcasts}}{show a list of broadcasts for the game(s)}
  \item{\code{linescore}}{show a linescore for completed game(s)}
  \item{\code{ticket}}{show different ticket purchase options for upcoming game(s)}
}}
}
\value{
The API output of schedules
}
\description{
Get the NHL schedule for a team, a date (range) or with additional information.
}
\examples{
#see the schedule:
schedule<-getSchedule()

#see the schedule linescores:
linescores<-getSchedule(scheduleType = 'linescore')
}
