% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{getSeasonInfo}
\alias{getSeasonInfo}
\title{Get Season Info}
\usage{
getSeasonInfo(season = NULL)
}
\arguments{
\item{season}{Optional, get specific season information. Season as 20172018 format, or 'current' for current season}
}
\value{
The API output of seasons
}
\description{
Get season information, such as start/end dates, Olympics participation, use of ties, etc.
}
\examples{
#See information on NHL Season 20172018:
season1718 <- getSeasonInfo(20172018)
#See the current season info
currentSeason <- getSeasonInfo('current')
}
