% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getSeasonSchedule}
\alias{getSeasonSchedule}
\title{Get Season Schedule}
\usage{
getSeasonSchedule(season)
}
\arguments{
\item{season}{Season's schedule to pull up.}
}
\value{
a data frame
}
\description{
This function returns a nicely formattd season schedule dataframe including home and away teams, date & start time, venue information, and game code.
}
\examples{
\dontrun{
#See the 20182019 schedule:
schedule <- getSeasonSchedule(20182019)

#See only preseason games from 20182019
schedule[schedule$gameType == 1, ]
}

}
