% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getSeasonScores}
\alias{getSeasonScores}
\title{Get Season Scores}
\usage{
getSeasonScores(season = NULL, schedule = NULL, progress_bar = TRUE)
}
\arguments{
\item{season}{Season's scores to get. Ignored if schedule provided, otherwise gets schedule from \code{getSeasonSchedule()}.}

\item{schedule}{Optional, if season provided. Season schedule from \code{getSeasonSchedule()}.}

\item{progress_bar}{Optional, logical, if progress bar as scores are retrieved.}
}
\value{
a data frame
}
\description{
Gets season scores (or scores for all games in a schedule), including OT & SO status, shots, goals, and shootout attempts & goals if applicable
}
\seealso{
getSeasonSchedule
}
