% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{getStandings}
\alias{getStandings}
\title{Get Standing}
\usage{
getStandings(standingsType = NULL, season = NULL, expand = FALSE)
}
\arguments{
\item{standingsType}{Optional, get specific standing type. See \code{\link{getStandingsTypes}()} for allowed values}

\item{season}{Get statndings for a specific season. If null, current season returned. Season should be in format 20172018. Overrides date specification.}

\item{expand}{Whether to return detailed information from each team.}
}
\value{
The API output of standings
}
\description{
Get the NHL standings, including historical standings. Or, get one single standing
}
\examples{
# get standings
standings <- getStandings()

#Get standings from 20182009
standings <- getStandings(season = 20082009)
}
