% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{getStandingsTypes}
\alias{getStandingsTypes}
\title{Get Standings Types}
\usage{
getStandingsTypes()
}
\value{
a list of standings types to call with \code{\link{getStandings}()}
}
\description{
Only certain standings display types are accepted. This returns the full valid list.
}
