% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{getTeam}
\alias{getTeam}
\title{Get Team}
\usage{
getTeam(team = NULL, modifier = NULL, season = NULL)
}
\arguments{
\item{team}{Optional, get specific team information.}

\item{modifier}{Optional, if team is supplied, get specific info about a team. Options are:
- 'person.names' Basic roster info.
- 'team.schedule.next' Returns details of the upcoming game for a team
- 'team.schedule.previous' Same as above but for the last game played}

\item{season}{Optional, if looking for season specific information on team}
}
\value{
The API output of teams
}
\description{
Get the NHL teams, including historical teams. Or, get one single team
}
