% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{getTeamLogos}
\alias{getTeamLogos}
\title{Get Team Logos}
\usage{
getTeamLogos()
}
\value{
a list of all team logo svg urls from the api
}
\description{
get Team Logo svg urls from NHL website
}
\examples{
logos <- getTeamLogos()
}
