% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournaments.R
\name{getTournament}
\alias{getTournament}
\title{Get Tournament}
\usage{
getTournament(season = NULL, expand = FALSE)
}
\arguments{
\item{season}{Optional, get specific season's playoff information}

\item{expand}{Optional, boolean, get expanded information}
}
\value{
The API output of tournaments/playoffs
}
\description{
Get the NHL tournaments (playoffs)
}
\examples{
#Get info on the 2017-2018 playoffs
playoff20172018<-getTournament(season = 20172018, expand = TRUE)
}
