% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awards.R
\name{getTrophy}
\alias{getTrophy}
\title{Get Trophys}
\usage{
getTrophy(name = NULL)
}
\arguments{
\item{name}{Optional, trophy name}
}
\value{
trophy information from records api. Has ability to search by name
}
\description{
Get Trophys
}
\examples{
#See list of trophys
trophies <- getTrophy()

#See only the Vezina trophy
vezina <- getTrophy('Vezina Trophy')
}
