% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venues.R
\name{getVenue}
\alias{getVenue}
\title{Get Venue}
\usage{
getVenue(venue = NULL)
}
\arguments{
\item{venue}{Optional, get specific venue information.}
}
\value{
The API output of venues
}
\description{
Get the NHL venues, or, get one single venue
}
