% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhlRapi.R
\name{print.nhl_api}
\alias{print.nhl_api}
\title{Overload print when S3 object generated}
\usage{
\method{print}{nhl_api}(x, ...)
}
\arguments{
\item{x}{nhl_api object}

\item{...}{}
}
\description{
Overload print when S3 object generated
}
