% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{querybuilder}
\alias{querybuilder}
\title{Query Builder}
\usage{
querybuilder(...)
}
\arguments{
\item{...}{all items to collapse to one query}
}
\value{
a string query of .../...
}
\description{
Generic wrapper to build querys using paste0 collapsing on '/'
}
\keyword{internal}
