% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhlRapi.R
\name{baseAPI}
\alias{baseAPI}
\title{Ultimate BASE api call.}
\usage{
baseAPI(call_url, query_wrapper)
}
\arguments{
\item{call_url}{the generated url to get}

\item{query_wrapper}{specific query to api endpoint}
}
\value{
json response
}
\description{
Ultimate BASE api call.
}
