% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{getAllTimeRecordVsFranchise}
\alias{getAllTimeRecordVsFranchise}
\title{Get Franchise Head to Head record}
\usage{
getAllTimeRecordVsFranchise(franchiseID = NULL, franchiseName = NULL,
  opponentName = NULL, opponentID = NULL)
}
\arguments{
\item{franchiseID}{Optional franchise ID to filter results.}

\item{franchiseName}{Optional franchise name to filter results.}

\item{opponentName}{Optional opponent name to filter results.}

\item{opponentID}{Optional opponent franchise ID to filter results.}
}
\value{
Head to head records for every franchise.
}
\description{
Get Franchise Head to Head record
}
