% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{getAttendance}
\alias{getAttendance}
\title{Get Attendance Records}
\usage{
getAttendance(season = NULL)
}
\arguments{
\item{season}{Optional, to filter attendance to a given season. Format: 20172018}
}
\value{
season attendance records from the API
}
\description{
Returns league-wide season attendance totals (no game-by-game or team-by-team breakdown available)
}
\examples{
#See all attendance records
getAttendance()

#Only see attendance from a certain year
getAttendance(20152016)
}
