% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{getFranchiseSeasonResults}
\alias{getFranchiseSeasonResults}
\title{Get Franchise Season Results Stats}
\usage{
getFranchiseSeasonResults(franchiseID = NULL, franchiseName = NULL,
  gameType = NULL, season = NULL)
}
\arguments{
\item{franchiseID}{Optional franchise ID to filter results.}

\item{franchiseName}{Optional franchise name to filter results.}

\item{gameType}{Optional, limit results to regular season (2) or playoffs (3). Accepts numerical input only}

\item{season}{Optional, limit results to a particular season}
}
\value{
Season record stats for every franchise.
}
\description{
Franchise season results
}
\examples{
#See the St. Louis Blues' regular season results
results <- getFranchiseSeasonResults(franchiseName = 'St. Louis Blues', gameType = 2)
}
