% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{getFranchiseSeasonResults}
\alias{getFranchiseSeasonResults}
\title{Get Franchise Season Results Stats}
\usage{
getFranchiseSeasonResults(franchiseID = NULL)
}
\arguments{
\item{franchiseID}{Optional franchise ID to filter results.}
}
\value{
Season record stats for every franchise.
}
\description{
Get Franchise Season Results Stats
}
