% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{getFranchiseSkaterRecords}
\alias{getFranchiseSkaterRecords}
\title{Get Franchise Season Record Stats}
\usage{
getFranchiseSkaterRecords(franchiseID = NULL)
}
\arguments{
\item{franchiseID}{Optional franchise ID to filter results.}
}
\value{
Season record stats for every franchise.
}
\description{
Get Franchise Season Record Stats
}
