% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{getGameBoxscore}
\alias{getGameBoxscore}
\title{Get Game Boxscore}
\usage{
getGameBoxscore(game)
}
\arguments{
\item{game}{Required, get specific game boxscore.}
}
\value{
The boxscore API output of games
}
\description{
Get a NHL game boxscore. Includes goals, shots, PIMs, blocked, takeaways, giveaways and hits
}
