% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{getGameFeed}
\alias{getGameFeed}
\title{Get Game Feed}
\usage{
getGameFeed(game, from = NULL)
}
\arguments{
\item{game}{Required, get specific game feed}

\item{from}{Time from which to update, format yyyymmdd_hhmmss}
}
\value{
The feed API output of games
}
\description{
Get a NHL game feed. Includes play by play data with event locations, and after-game info like 3 stars.
}
