% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestones.R
\name{getMilestone}
\alias{getMilestone}
\title{Get Milestones Records}
\usage{
getMilestone(milestone)
}
\arguments{
\item{milestone}{Required, one of:
- 1000-point-career
- 500-goal-career
- 100-point-season
- 50-goal-season
- 5-goal-game}
}
\value{
milestones from the API
}
\description{
Get Milestones Records
}
