% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{getPlayersByTeam}
\alias{getPlayersByTeam}
\title{Get Players per team}
\usage{
getPlayersByTeam(team)
}
\arguments{
\item{team}{required, team id}
}
\value{
the players (from records api) on the team
}
\description{
This uses a different API endpoint from the other team gets. Pulls players on a team currently.
}
\examples{
#See the active players on Edmonton Oilers (ID = 22)
players <- getPlayersByTeam(22)
}
