% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{getPlayerStats}
\alias{getPlayerStats}
\title{Get Player's Stats}
\usage{
getPlayerStats(player, stat = NULL, season = NULL)
}
\arguments{
\item{player}{The player ID to get stats}

\item{stat}{The speific stats to get. Many options exist, see \code{\link{getPlayerStatTypes}()} for full list}

\item{season}{a season, if required or desired, for querying stats. Format: 20172018 for the 2017-2018 season}
}
\value{
The API output of player
}
\description{
get specific stats for a provided player
}
