% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{getRecordsList}
\alias{getRecordsList}
\title{Get Records List}
\usage{
getRecordsList()
}
\value{
a list of all records maintained by the NHL. The return is the API URL to call for the current record result.
}
\description{
Get Records List
}
\examples{
records<-getRecordsList()
}
