% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{getTeam}
\alias{getTeam}
\title{Get Team}
\usage{
getTeam(team = NULL, modifier = NULL)
}
\arguments{
\item{team}{Optional, get specific team information.}

\item{modifier}{Optional, if team is supplied, get specific info about a team. Options are:
- 'team.roster' Shows roster of active players for the specified team
- 'person.names' Same as above, but gives less info.
- 'team.schedule.next' Returns details of the upcoming game for a team
- 'team.schedule.previous' Same as above but for the last game played
- 'team.stats' Returns the teams stats for the season
- list('team.roster', 'season=20142015') Adding the season identifier shows the roster for that season
- list('team.stats', 'stats=statsSingleSeasonPlayoffs') Specify which stats to get.}
}
\value{
The API output of teams
}
\description{
Get the NHL teams, including historical teams. Or, get one single team
}
