% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{getTeamRoster}
\alias{getTeamRoster}
\title{Get Team Roster}
\usage{
getTeamRoster(team, season = NULL)
}
\arguments{
\item{team}{The team roster to get. Required}

\item{season}{Which season's roster to get. Optional, will get the current roster if not supplied.}
}
\value{
the API output of teams' rosters
}
\description{
Returns the team roster, configurable for season. Convenience wrapper around \code{\link{getTeam}(team = team, modifier = team.roster, season = season)}.
}
\examples{
#See the Carolina Hurricane's roster (ID = 12)
canes <- getTeamRoster(12)

#Get their roster a few years ago
canes_past <- getTeamRoster(12, season = 20152016)
}
