% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{getTeamStats}
\alias{getTeamStats}
\title{Get Team Stats}
\usage{
getTeamStats(team, season = NULL)
}
\arguments{
\item{season}{Which season's stats to get. Optional, will get the current years' stats if not supplied}

\item{teamThe}{team stats to get. Required.}
}
\value{
Team stats from the API
}
\description{
Returns the team stats, both single season numbers and league ranking. Convenience wrapper around \code{\link{getTeam}(team = team, modifier = team.stats, season = season)}.
}
